mngr e2e: reconcile tutorial tests with current mega_tutorial.sh
Resolve the 11 pytest functions that the tutorial_matcher reported as
unmatched after the e2e split:

- Drop 5 duplicates from tutorial/test_basic.py whose canonical
  versions already live in tutorial/test_list.py, test_create_basic.py,
  test_destroy.py, and test_rename.py.
- Refresh stale write_tutorial_block() arguments in
  tutorial/test_create_basic.py::test_create_in_place,
  tutorial/test_create_modal.py::test_create_provider_modal, and
  tutorial/test_create_modal.py::test_create_modal_upload_and_extra_provision_command
  to match the current tutorial wording.
- Remove tutorial/test_create_modal.py::test_create_modal_retry (the
  retry block was reduced to a comment-only section, leaving no
  executable command for the matcher to anchor on).
- Remove tutorial/test_create_data_and_git.py::test_create_with_transfer_git_mirror
  (the --transfer=git-mirror block was removed from the tutorial).
- Move tutorial/test_rename.py to top-level e2e/test_rename.py since
  the tutorial's RENAMING AGENTS section is now informational only; the
  rename test stays as a non-tutorial e2e test.

Also move two tests that did correspond to tutorial blocks but lived at
the top level into tutorial/, attaching their write_tutorial_block:
- test_multi_agent.py::test_destroy_all_via_stdin -> tutorial/test_destroy.py
- test_plugin.py::test_plugin_list_shows_installed -> new
  tutorial/test_plugin.py
